Update Using Kamal with Ubicloud to work with Rails 8 and Kamal 2
The previous instructions were for Rails 7 and Kamal 1, I think.

This drops the use of the load balancer, and uses only a single
VM instead of multiple VMs.  One reason for this is I think we
should avoid unnecessary complexity in this guide, which is aimed
at new users.  The other reason is I have no idea how to integrate
Ubicloud load balancers with Kamal 2.